bug fix and test

+++ src/frame.py
@@ -7,8 +7,8 @@ class Frame:
     def __init__(self, cfg):
         self.h = copy(cfg["image_height"])
         self.w = copy(cfg["image_width"])
-        self.image = np.zeros((self.h, self.w, 3))
-        self.label = np.zeros((self.h, self.w, 3))
+        self.image = np.zeros((self.h, self.w, 3), dtype=np.uint8)
+        self.label = np.zeros((self.h, self.w, 3), dtype=np.uint8)
 
     def write(self, frame_image_path, frame_label_path):
         i = Image.fromarray(self.image)
